Rename buildEpisodeCoverageString.ts to releaseManagementHelpers.ts
