+++ src/core/utilities/releaseManagementHelpers.ts
